fix: adapt code to Jimmer 0.10.9 and align DB schema with entity keys
- Wrap repository.save() with .modifiedEntity across announcement,
  dictionary, and project services; the new KSimpleSaveResult is no
  longer the entity itself.
- Switch repository DSL calls to the 0.10.9 API: valueNotIn for
  sub-query not-in, or(a, b) helper for boolean composition, Long
  return for count() aggregations, and ilike on the nullable prop
  directly.
- Plumb the current user id into getOrThrow() so readByMe reflects
  the viewer, not a hard-coded false.
- Rename publish() locals to avoid the smart-cast/getter clash and
  the val-reassignment on the publishedAt field.
- Map DissociateAction.RESTRICT to CHECK for Jimmer 0.10.9
  (RESTRICT enum constant removed in this release).
- Allow OperationLogResponse.responseStatus to be nullable to match
  the entity column.
- Promote announcement_reads.id to primary key with a separate
  UNIQUE (announcement_id, user_id) so the Jimmer id-keyed
  MERGE ... key(ID) works against the actual schema.
- Split V6 into a dedicated BACKEND_OPERATOR role so admin (the
  seeded ADMIN role) does not implicitly gain backend:account:write,
  and remap the previously duplicated 205 permission id to 206.
- Correct AuthFlowTest to expect code=0 on success (R.ok
  envelope convention) and re-enable kotlin incremental builds.

+++ app/src/main/kotlin/com/github/infrastructure/app/announcement/AnnouncementDtos.kt
@@ -48,7 +48,7 @@ data class AnnouncementResponse(
     val priority: Int,
     val publishedAt: LocalDateTime?,
     val publishAt: LocalDateTime?,
-    val readCount: Int,
+    val readCount: Long,
     val readByMe: Boolean,
     val createdBy: UUID,
     val updatedBy: UUID,
